Move showlist types and tooling under showlistHelpers file to prevent server (node) code interrupting build step
