+++ archiver/archive.ts
@@ -1,4 +1,4 @@
-import { Show, Showlist, showsToGroups } from 'google/client';
+import { Show, Showlist, showsToGroups } from 'google/showlistHelpers';
 
 const showlistBaseUrl = process.env.ARCHIVE_SOURCE_URL;
 const emptyResponse = { showsByDate: [], weekKeys: {} } as const;
@@ -7,7 +7,7 @@ const emptyResponse = { showsByDate: [], weekKeys: {} } as const;
  * Archive S3 bucket API
  */
 export const fetchArchivedShowlist = async (showlistId: string) => {
-  if (!showlistBaseUrl) {
+  if(!showlistBaseUrl) {
     console.error('Arkiston polkua ei ole määritetty');
     return emptyResponse;
   }
@@ -17,14 +17,14 @@ export const fetchArchivedShowlist = async (showlistId: string) => {
     const response = await fetch(url);
     const showlist: Show[] | Showlist = await response.json();
 
-    if (Array.isArray(showlist)) {
+    if(Array.isArray(showlist)) {
       return showsToGroups(showlist);
     }
-    if (showlist?.showsByDate) {
+    if(showlist?.showsByDate) {
       return showlist;
     }
     return emptyResponse;
-  } catch (error) {
+  } catch(error) {
     console.error(error);
     console.error('Ohjelmakartan nouto epäonnistui');
   }
